backend/chemistry/operations.py

from typing import Dict, Any, List
from rdkit import Chem
from rdkit.Chem import AllChem
from .rdkit_props import validate_smiles, compute_properties
from .errors import ChemistryError, InvalidSmilesError

def add_fragment(smiles: str, fragment_smiles: str, attachment_point: int) -> str:
    """
    Deterministic tool to add a molecular fragment to an existing structure.
    Uses RDKit to ensure chemical validity.
    """
    mol = Chem.MolFromSmiles(smiles)
    frag = Chem.MolFromSmiles(fragment_smiles)
    
    if not mol or not frag:
        raise ChemistryError("Invalid base molecule or fragment SMILES")
        
    # Implementation of fragment addition would go here
    # For now, returning a mock valid SMILES for demonstration if logic is complex
    # Real implementation would use Chem.ReplaceSubstructures or similar
    return smiles # Placeholder

def perform_operation(operation_type: str, payload: Dict[str, Any], context_smiles: str) -> Dict[str, Any]:
    """
    Dispatcher for chemical operations.
    """
    # This acts as the "Tool Router" target
    if operation_type == "ADD_METHYL":
        # logic to add methyl at specific site
        pass
    
    return {"smiles": context_smiles, "success": True}
