backend/requirements.txt
##############################
# Core Framework
##############################
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.9.2
python-dotenv==1.0.1

##############################
# Chemistry / Molecules
##############################
# Note: RDKit is included in base image, but keeping for version pinning
rdkit-pypi==2022.9.5
pillow==10.4.0

##############################
# Backend & DB
##############################
sqlmodel==0.0.27
httpx==0.27.0
python-multipart==0.0.6
redis==5.0.1
rq==1.15.1

##############################
# ML/AI (lightweight - heavy deps in requirements-model.txt)
##############################
# Or use requirements-model.txt for full ML stack
google-generativeai==0.8.6

##############################
# Testing
##############################
pytest==7.4.2